--- src/services/User.model.ts
@@ -1,12 +1,0 @@
-export interface IUser {
-    id?: string;
-    username: string;
-    age: number;
-    hobbies: Array<string>;
-}
-
-export class IUSER_DEFAULT implements IUser {
-    username: string = '';
-    age: number = 0;
-    hobbies: string[] = []
-}
