+++ src/utils/board.validate.ts
@@ -1,4 +1,4 @@
-import { BadRequestError, BoardBadRequestError, BoardValidationError } from "../Errors/CustomErrors";
+import { BoardBadRequestError, BoardValidationError } from "../Errors/CustomErrors";
 import { IBoard } from "../services/board/Board.model";
 
 export const boardValidate = (board: IBoard) => {
